More MP changes
