+++ microprofile/openapi/etc/spotbugs/exclude.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
 
-    Copyright (c) 2019 Oracle and/or its affiliates.
+    Copyright (c) 2019, 2023 Oracle and/or its affiliates.
 
     Licensed under the Apache License, Version 2.0 (the "License");
     you may not use this file except in compliance with the License.
@@ -30,5 +30,14 @@
         <Class name="io.helidon.microprofile.openapi.IndexBuilder"/>
         <Method name="checkForIndexFile" />
         <Bug pattern="OBL_UNSATISFIED_OBLIGATION" />
+    </Match>
+
+    <!-- We need the following due to SpotBugs issue 1877
+         https://github.com/spotbugs/spotbugs/issues/1877 (false positive with text blocks)
+    -->
+    <Match>
+        <Class name="io.helidon.microprofile.openapi.MPOpenAPIBuilder"/>
+        <Method name="filteredIndexView" />
+        <Bug pattern="VA_FORMAT_STRING_USES_NEWLINE" />
     </Match>
 </FindBugsFilter>
